// temp-generated/model/authChangePasswordPutRequest.ts
/**
 * E-Commerce Jewelry Store API
 * REST API for a jewelry e-commerce application
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';

export class AuthChangePasswordPutRequest {
    'old_password': string;
    'new_password': string;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "old_password",
            "baseName": "old_password",
            "type": "string"
        },
        {
            "name": "new_password",
            "baseName": "new_password",
            "type": "string"
        }    ];

    static getAttributeTypeMap() {
        return AuthChangePasswordPutRequest.attributeTypeMap;
    }
}