// temp-generated/model/authSession.ts
/**
 * E-Commerce Jewelry Store API
 * REST API for a jewelry e-commerce application
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';
import { User } from './user';

export class AuthSession {
    'user'?: User;
    /**
    * JWT token for subsequent authenticated requests.
    */
    'token'?: string;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "user",
            "baseName": "user",
            "type": "User"
        },
        {
            "name": "token",
            "baseName": "token",
            "type": "string"
        }    ];

    static getAttributeTypeMap() {
        return AuthSession.attributeTypeMap;
    }
}