// temp-generated/model/cartCheckoutPost200Response.ts
/**
 * E-Commerce Jewelry Store API
 * REST API for a jewelry e-commerce application
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';

export class CartCheckoutPost200Response {
    /**
    * URL for the user to complete payment.
    */
    'session_url'?: string;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "session_url",
            "baseName": "session_url",
            "type": "string"
        }    ];

    static getAttributeTypeMap() {
        return CartCheckoutPost200Response.attributeTypeMap;
    }
}