// temp-generated/model/cartInsert.ts
/**
 * E-Commerce Jewelry Store API
 * REST API for a jewelry e-commerce application
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';

export class CartInsert {
    'item_id': number;
    'quantity': number;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "item_id",
            "baseName": "item_id",
            "type": "number"
        },
        {
            "name": "quantity",
            "baseName": "quantity",
            "type": "number"
        }    ];

    static getAttributeTypeMap() {
        return CartInsert.attributeTypeMap;
    }
}