// temp-generated/model/cartItem.ts
/**
 * E-Commerce Jewelry Store API
 * REST API for a jewelry e-commerce application
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';
import { ItemFetch } from './itemFetch';

export class CartItem {
    'item_id': number;
    'quantity': number;
    'added_at'?: Date;
    'item_details'?: ItemFetch;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "item_id",
            "baseName": "item_id",
            "type": "number"
        },
        {
            "name": "quantity",
            "baseName": "quantity",
            "type": "number"
        },
        {
            "name": "added_at",
            "baseName": "added_at",
            "type": "Date"
        },
        {
            "name": "item_details",
            "baseName": "item_details",
            "type": "ItemFetch"
        }    ];

    static getAttributeTypeMap() {
        return CartItem.attributeTypeMap;
    }
}