// temp-generated/model/earringDetails.ts
/**
 * E-Commerce Jewelry Store API
 * REST API for a jewelry e-commerce application
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';
import { Colors } from './colors';

export class EarringDetails {
    /**
    * Size dimension (e.g., diameter, length in mm).
    */
    'earring_size': number;
    'earring_color': Colors;
    /**
    * Specific style of the earring (e.g., Stud, Hoop, Dangle).
    */
    'earring_style': string;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "earring_size",
            "baseName": "earring_size",
            "type": "number"
        },
        {
            "name": "earring_color",
            "baseName": "earring_color",
            "type": "Colors"
        },
        {
            "name": "earring_style",
            "baseName": "earring_style",
            "type": "string"
        }    ];

    static getAttributeTypeMap() {
        return EarringDetails.attributeTypeMap;
    }
}

export namespace EarringDetails {
}