// temp-generated/model/itemBase.ts
/**
 * E-Commerce Jewelry Store API
 * REST API for a jewelry e-commerce application
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';
import { GemShape } from './gemShape';
import { GemSize } from './gemSize';
import { JewelryType } from './jewelryType';

export class ItemBase {
    /**
    * Unique identifier for the item.
    */
    'id': number;
    'gem_size': GemSize;
    'gem_shape': GemShape;
    /**
    * Name of the gemstone.
    */
    'gem_name': string;
    'jewelry_type': JewelryType;
    /**
    * Auto-generated display name (e.g., \'Diamond Ring\').
    */
    'item_name': string;
    /**
    * Price of the item.
    */
    'price': number;
    /**
    * Indicates if the item is currently in stock.
    */
    'in_stock': boolean;
    /**
    * URL of the item\'s image.
    */
    'img_url': string;
    /**
    * Timestamp when the item was created.
    */
    'created_at': Date;
    /**
    * Timestamp when the item was last updated.
    */
    'updated_at': Date;
    /**
    * Indicates if the item is featured.
    */
    'is_featured': boolean;
    /**
    * Indicates if the item is hidden from public view.
    */
    'is_hidden': boolean;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "id",
            "baseName": "id",
            "type": "number"
        },
        {
            "name": "gem_size",
            "baseName": "gem_size",
            "type": "GemSize"
        },
        {
            "name": "gem_shape",
            "baseName": "gem_shape",
            "type": "GemShape"
        },
        {
            "name": "gem_name",
            "baseName": "gem_name",
            "type": "string"
        },
        {
            "name": "jewelry_type",
            "baseName": "jewelry_type",
            "type": "JewelryType"
        },
        {
            "name": "item_name",
            "baseName": "item_name",
            "type": "string"
        },
        {
            "name": "price",
            "baseName": "price",
            "type": "number"
        },
        {
            "name": "in_stock",
            "baseName": "in_stock",
            "type": "boolean"
        },
        {
            "name": "img_url",
            "baseName": "img_url",
            "type": "string"
        },
        {
            "name": "created_at",
            "baseName": "created_at",
            "type": "Date"
        },
        {
            "name": "updated_at",
            "baseName": "updated_at",
            "type": "Date"
        },
        {
            "name": "is_featured",
            "baseName": "is_featured",
            "type": "boolean"
        },
        {
            "name": "is_hidden",
            "baseName": "is_hidden",
            "type": "boolean"
        }    ];

    static getAttributeTypeMap() {
        return ItemBase.attributeTypeMap;
    }
}

export namespace ItemBase {
}